Update VM initialization script with latest kernel requirements
The swystem kernel fork (which is up to date with the latest kernel
version), has different dependencies compared to last year, when the
setup guide was created.
Specifically, the kernel requires a more recent rust-bindgen version in
order to have the Rust support.
Note that this file is still temporary, and the actual virtual machine
setup should be done using Ansible. I still need to check it out
properly, so for now this script works well enough.